Add missing casts to satisfy -Wc++11-narrowing

+++ test_common/harness/rounding_mode.c
@@ -132,7 +132,7 @@ void *FlushToZero( void )
 {
 #if defined( __APPLE__ ) || defined(__linux__) || defined (_WIN32)
     #if defined( __i386__ ) || defined( __x86_64__ ) || defined(_MSC_VER)
-        union{ int i;  void *p; }u = { _mm_getcsr() };
+        union{ int i;  void *p; }u = { (int)_mm_getcsr() };
         _mm_setcsr( u.i | 0x8040 );
         return u.p;
     #elif defined( __arm__ )
