+++ test_conformance/basic/test_imagecopy3d.c
@@ -179,7 +179,7 @@ test_imagecopy3d(cl_device_id device, cl_context context, cl_command_queue queue
 				break;
 		}
 		
-        size_t origin[3]={0,0,0}, region[3]={img_width, img_height, img_depth};
+        size_t origin[3]={0,0,0}, region[3]={static_cast<size_t>(img_width), static_cast<size_t>(img_height), static_cast<size_t>(img_depth)};
         err = clEnqueueWriteImage(queue, streams[i*2], CL_TRUE, origin, region, 0, 0, p, 0, NULL, NULL);
         test_error(err, "clEnqueueWriteImage failed");
