+++ test_conformance/basic/test_writeimage.c
@@ -124,7 +124,7 @@ int test_writeimage(cl_device_id device, cl_context context, cl_command_queue qu
 	int img_height = 512;
 	int i, err, any_err = 0;
     size_t origin[3] = {0, 0, 0};
-    size_t region[3] = {img_width, img_height, 1};
+    size_t region[3] = {static_cast<size_t>(img_width), static_cast<size_t>(img_height), 1};
     size_t length = img_width * img_height * 4 * sizeof(unsigned char);
 
 	PASSIVE_REQUIRE_IMAGE_SUPPORT( device )
